+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.13.1
 
 - Added sort by closed date option in task list view and emedded task lists.
+- Added search filters but state groups `state:active`, `state:completed`, `state:inactive`, and `state:waiting`.
 - Fixed issue with active tasks being struckthrough in embedded task list when extended markdown checkbox option enabled.
 
 ## 0.13.0